backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.8.0
python-dotenv==1.0.1
bcrypt==4.2.0
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
gunicorn==22.0.0
